requirements.txt
alabaster==1.0.0
anyio==4.10.0
asgiref==3.9.1
babel==2.17.0
black==25.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
Django==5.2.5
django-environ==0.12.0
djangorestframework==3.16.1
docutils==0.21.2
h11==0.16.0
idna==3.10
imagesize==1.4.1
isort==6.0.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mypy_extensions==1.1.0
mysqlclient==2.2.7
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
Pygments==2.19.2
python-dotenv==1.1.1
requests==2.32.5
requests-oauthlib==2.0.0
roman-numerals-py==3.1.0
ruff==0.13.0
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinx-autobuild==2025.8.25
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
starlette==0.47.3
tweepy==4.16.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
